fix(layout): Move Release & translation fields into dedicated Release section
Extract newsReleaseOriginId, newsReleaseDistributionId, translationsRequiredStatusId, and translationLanguageIds from the Comms section into a new ActivityReleaseSection.
Rename section label newsRelease from 'News release' to 'Release'.
Render ActivityReleaseSection above ActivityEventSection in the shared form body and pass required lookups.
Clean up ActivityCommsSection props/variables (removed release & translations responsibilities).
Export the new section and update imports.
Ran full build and test suite — build/tests pass.
Files changed (key):

ActivityReleaseSection.tsx
ActivityCommsSection.tsx
ActivityFormBody.tsx
index.ts
activity-form-section-labels.ts
