+++ calendar-ui/src/components/activity/ActivityFormSections/index.ts
@@ -1,6 +1,7 @@
 export { ActivityOverviewSection } from './ActivityOverviewSection';
 export { ActivityScheduleSection } from './ActivityScheduleSection';
 export { ActivityCommsSection } from './ActivityCommsSection';
+export { ActivityReleaseSection } from './ActivityReleaseSection';
 export { ActivityEventSection } from './ActivityEventSection';
 export { ActivityReportsSection } from './ActivityReportsSection';
 export { ActivitySharingSection } from './ActivitySharingSection';
